% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_wave.R
\name{allocate_wave}
\alias{allocate_wave}
\title{Adaptive Multi-Wave Sampling}
\usage{
allocate_wave(data, strata, y, wave2a, n)
}
\arguments{
\item{data}{A data frame or matrix with one row for each sampling unit, one column specifying each unit's stratum, one column holding the value of the continuous variable for which the variance should be minimized, and one column containing a key specifying if each unit has already been sampled.}

\item{strata}{a character string or vector of character strings specifying the name of columns which indicate the stratum that each unit belongs to.}

\item{y}{a character string specifying the name of the continuous variable for which the variance should be minimized.}

\item{wave2a}{a character string specifying the name of a column that contains a binary (Y/N or 1/0) indicator specifying whether each unit has already been sampled in a previous wave.}

\item{n}{The desired sample size of the next wave.}
}
\value{
Returns a dataframe with the n allocated to each strata for the next sampling wave.
}
\description{
Determines the optimal sampling allocation for a new sampling wave based on results from previous waves.
}
