% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_simdata.R
\docType{data}
\name{example_simdata}
\alias{example_simdata}
\alias{Example_SimPhase1}
\alias{Example_SimPhase2a}
\alias{Example_SimPhase2b}
\title{Example Dataset: Maternal Weights}
\format{
\code{Phase1:} a data frame with 10335 rows and 7 columns
\describe{
 \item{\code{id}}{unique ID for each mother-child pair}
 \item{\code{event}}{binary indicator for event of child obesity}
 \item{\code{mother_weight_change_est}}{estimated mother weight change during pregnancy}
 \item{\code{race}}{race of child}
 \item{\code{ttobesity}}{follow up time after childbirth in years}
 \item{\code{obesity}}{binary indicator for diabetes in the mother}
 \item{\code{mat_age}}{mother's age at childbirth}
 }

\code{Phase2a:} a data frame with 252 rows and 6 columns.
\describe{
 \item{\code{child_age}}{validated time of follow up in years after childbirth}
 \item{\code{nchildren}}{birth order of child}
 \item{\code{wgt_change_validated}}{validated weight change of mother during pregnancy}
 \item{\code{mat_obesity}}{binary indicator for maternal obesity}
 \item{\code{mat_race}}{mother's race}
 \item{\code{delivery_method}}{birth delivery method}
}

\code{Phase2b:} a data frame with 248 rows and 6 columns. The variable names are the same as in Phase2a.
}
\usage{
Example_SimPhase1

Example_SimPhase2a

Example_SimPhase2b
}
\description{
These three *simulated* datasets contain data on demographic characteristics and clinical data related to childhood obesity for 10335 mother-child pairs. They are used to generate the workflow in the main package vignette. They are based on a study that used multi-wave adaptive sampling to validate electronic health records that target factors related to childhood obesity (see https://www.pcori.org/research-results/2017/developing-methods-estimate-and-address-errors-studies-using-electronic-health).
}
\details{
The Phase 1 dataset contains unvalidated data on 10,335 mother-child pairs. The Phase 2a and 2b datasets contain validated data on a sample 252 and 248 pairs respectively. See package vignettes for more details.
}
\keyword{datasets}
