% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwave.R
\docType{class}
\name{multiwave-class}
\alias{multiwave-class}
\alias{multiwave}
\title{Multiwave Class for Multi-Wave Sampling Organization}
\description{
\pkg{optimall} defines three S4 classes for organizing the
multi-wave sampling workflow: \code{wave}, \code{phase}, and \code{multiwave}.
An object of class \code{multiwave} holds metadata and a list of objects of
class \code{phase}, which in turn holds metadata and a list of objects of class
\code{wave}. These three object classes are used together to organize the
workflow of multi-wave sampling designs.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list of elements that describe the entire survey.
The list is empty upon initialization of the multiwave object, but the user
may add anything to it as they see fit. It may include a "title".}

\item{\code{phases}}{A list of objects of class \code{phase} (see other class
documentation).}
}}

