% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_allocation.R
\name{optimal_allocation}
\alias{optimal_allocation}
\title{Optimal Allocation}
\usage{
optimal_allocation(
  data,
  strata,
  y,
  nsample = NULL,
  ndigits = 2,
  method = "WrightII",
  allow.na = FALSE
)
}
\arguments{
\item{data}{A data frame or matrix with one row for each sampled unit, one column specifying each unit's stratum, and one column holding the value of the continuous variable for which the variance should be minimized.}

\item{strata}{a character string or vector of character strings specifying the name of columns which indicate the stratum that each unit belongs to.}

\item{y}{a character string specifying the name of the continuous variable for which the variance should be minimized.}

\item{nsample}{the desired total sample size. Defaults to NULL.}

\item{ndigits}{a numeric value specifying the number of digits to round the stratum fraction to.}

\item{method}{a character string specifying the method of optimal sample allocation to use. Must be one of "Neyman", "WrightI" or "WrightII". Defaults to "WrightII".}

\item{allow.na}{logical input specifying whether y should be allowed to have NA values. Defaults to FALSE.}
}
\value{
Returns a data frame with the n allocated to each strata or the sampling fractions if nsample is NULL.
}
\description{
Determines the optimal sampling fraction and sample size for each stratum in a stratified random sample according to Neyman Allocation or Exact Optimum Sample Allocation (Wright 2014).
}
\examples{
optimal_allocation(data = iris, strata = "Species", y = "Sepal.Length",
nsample = 100, method = "WrightII")
}
\references{
Wright, T. (2014). A simple method of exact optimal sample allocation under stratification with any mixed constraint patterns. Statistics, 07.
}
