% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimall_shiny.R
\name{optimall_shiny}
\alias{optimall_shiny}
\title{Visualize Strata}
\usage{
optimall_shiny()
}
\description{
Launches an R Shiny application locally. This app can be used to
interactively split strata and determine how the results affect
optimum allocation of a fixed number of samples.
}
