% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimall_shiny.R
\name{optimall_shiny_run}
\alias{optimall_shiny_run}
\title{Run the shiny application}
\usage{
optimall_shiny_run(...)
}
\arguments{
\item{...}{Optional arguments to pass to \code{shiny::runApp}.
\code{display.mode} is already set to normal.}
}
\value{
Launches an R Shiny application locally.
}
\description{
Launches an R Shiny application locally. This app can be used to
interactively split strata and determine how the results affect
optimum allocation of a fixed number of samples.
}
