% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimum_allocation.data.frame.R
\name{optimum_allocation.data.frame}
\alias{optimum_allocation.data.frame}
\title{Method for class 'data.frame'}
\usage{
optimum_allocation.data.frame(
  data,
  strata,
  y,
  nsample = NULL,
  ndigits = 2,
  method = "WrightII",
  allow.na = FALSE,
  phase,
  wave
)
}
\description{
Method for class 'data.frame'
}
