% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\docType{class}
\name{phase-class}
\alias{phase-class}
\alias{phase}
\title{Phase Class for Multi-Wave Sampling Organization}
\description{
\pkg{optimall} defines three S4 classes for organizing the
multi-wave sampling workflow: \code{wave}, \code{phase}, and \code{multiwave}.
An object of class \code{multiwave} holds metadata and a list of objects of
class \code{phase}, which in turn holds metadata and a list of objects of class
\code{wave}. These three object classes are used together to organize the
workflow of multi-wave sampling designs.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list containing the phase metadata}

\item{\code{waves}}{A list of objects of class \code{wave}, each element
representing one wave of the phase}
}}

