% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strata.data.frame.R
\name{sample_strata.data.frame}
\alias{sample_strata.data.frame}
\title{Generate IDs to sample in a dataframe}
\usage{
\method{sample_strata}{data.frame}(
  data1,
  strata1,
  id,
  wave2a = NULL,
  data2,
  strata2 = "strata",
  n_allocated = "n_to_sample",
  phase,
  wave
)
}
\description{
Generate IDs to sample in a dataframe
}
