% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strata.Multiwave.R
\name{sample_strata.Multiwave}
\alias{sample_strata.Multiwave}
\title{Generate IDs to sample in Multiwave Object}
\usage{
sample_strata.Multiwave(
  data1,
  phase,
  wave,
  strata1 = NULL,
  id = NULL,
  wave2a = NULL,
  data2 = NULL,
  strata2 = NULL,
  n_allocated = NULL
)
}
\description{
Generate IDs to sample in Multiwave Object
}
