% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\alias{set_data<-}
\alias{set_data<-,Multiwave-method}
\title{Write Slots of a Multiwave Object}
\usage{
set_data(
  x,
  phase = 1,
  wave = NA,
  slot = c("data", "design", "metadata", "samples", "sampled_data")
) <- value
}
\arguments{
\item{x}{an object of class \code{'Multiwave'}}

\item{phase}{a numeric value specifying the phase that should be accessed.
To access the overall metadata, set \code{phase = NA}. Defaults to 1.}

\item{wave}{a numeric value specifying the wave that should be accessed.
Ta access phase metadata, set \code{wave = NA}. Defaults to \code{NA}.}

\item{slot}{a character value specifying the name of the slot to be
accessed. Must be one of \code{"metadata"}, \code{"design"},
\code{"samples"}, \code{"sampled_data"}, \code{"data"}. Defaults to
\code{"data"}. See class documentation or package vignettes for more
information about slots.}

\item{value}{value to assign to specified slot}
}
\description{
\code{set_data} is used to assign values (write to) slots of
\code{Multiwave} class objects.
}
\section{Functions}{
\itemize{
\item \code{set_data(
  x,
  phase = 1,
  wave = NA,
  slot = c("data", "design", "metadata", "samples", "sampled_data")
) <- value}: assign value to slot of a multiwave object

}}
\examples{
# To write overall metadata
get_data(MySurvey, phase = NA, slot = "metadata") <- list(
  title = "Maternal Weight Survey"

# To write Phase 2 metadata
get_data(MySurvey, phase = 2, slot = "metadata") <- list(
  strata = "mystrata", id = "id")

}
