% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimall_shiny.R
\name{shiny_ui}
\alias{shiny_ui}
\title{UI for Shiny App for Splitting Strata with Optimum Allocation}
\usage{
shiny_ui()
}
\value{
Creates the UI for the Shiny app that is loaded with
\code{optimall_shiny}.
}
\description{
UI for Shiny App for Splitting Strata with Optimum Allocation
}
