% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_strata.R
\name{split_strata}
\alias{split_strata}
\title{Split Strata}
\usage{
split_strata(data, y, strata, split, split_perc = 0.5)
}
\arguments{
\item{strata}{a character string or vector of character strings specifying the columns that are being used to create the strata, leaving out the column that should be split.}

\item{split}{a character string specifying the column to split on.}
}
\description{
Splits strata
}
