% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multiwave.R
\name{summary,Multiwave-method}
\alias{summary,Multiwave-method}
\title{Method for summary for class Multiwave}
\usage{
\S4method{summary}{Multiwave}(object)
}
\arguments{
\item{object}{object of class "Multiwave"}
}
\description{
Method for summary for class Multiwave
}
