% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\docType{class}
\name{Wave-class}
\alias{Wave-class}
\alias{Wave}
\title{Wave Class for Multi-Wave Sampling Organization}
\description{
\pkg{optimall} defines three S4 classes for organizing the
multi-wave sampling workflow: \code{Wave}, \code{Phase},
and \code{Multiwave}.
An object of class \code{Multiwave} holds metadata and a list of objects of
class \code{Phase}, which in turn holds metadata and a list of
objects of class
\code{Wave}. These three object classes are used together to organize the
workflow of multi-wave sampling designs.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list containing the metadata for the wave.}

\item{\code{design}}{a dataframe specifying the design of the wave.
Is often the output of \code{allocate_wave}.}

\item{\code{samples}}{A character vector containing the ids of the units
sampled in the wave.}

\item{\code{sampled_data}}{A dataframe holding the data, with ids,
\emph{collected} in this wave of sampling}

\item{\code{data}}{A dataframe holding the updated full data set with all of
the Phase 1 sampling units including the samples collected in this wave.}
}}

