% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_wave.data.frame.R
\name{allocate_wave.data.frame}
\alias{allocate_wave.data.frame}
\title{allocate_wave data frame}
\usage{
\method{allocate_wave}{data.frame}(
  data,
  strata,
  y,
  wave2a,
  nsample,
  method = "iterative",
  detailed = FALSE
)
}
\description{
allocate_wave data frame
}
